# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Dispatch API
  version: 3.0.0
  description: |
    Open-source bulk email campaign platform with tracking, multi-provider support, and batch processing.

    ## Authentication
    Most endpoints require authentication via a `session_token` cookie obtained from the login endpoint.
    Alternatively, an API key can be passed via the `X-API-Key` header.

    ## CSRF Protection
    All mutation endpoints (POST, PUT, DELETE) require a valid `X-CSRF-Token` header.
    The CSRF token is issued as a cookie on every response.

    ## Common Response Format
    All endpoints return a standard envelope:
    ```json
    {
      "success": true,
      "data": { ... },
      "message": "Human-readable message",
      "meta": { "pagination": { ... } }
    }
    ```
  contact:
    name: Dispatch
    url: https://github.com/iamjpsingh/bulk-email-sender
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
      description: Session token cookie set on login/register.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for programmatic access.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF token required for POST/PUT/DELETE requests. Value is read from the `csrf_token` cookie.

  # -------------------------------------------------------------------------
  # Reusable Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          description: Response payload (shape varies per endpoint).
        message:
          type: string
        meta:
          type: object
          properties:
            pagination:
              $ref: '#/components/schemas/Pagination'
      required: [success]
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Something went wrong
        error:
          type: string
      required: [success, message]
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5
        hasMore:
          type: boolean
          example: true

    # -- Auth --
...
    SMTPConfigCreate:
      type: object
      properties:
        name:
          type: string
          default: Default Configuration
        host:
          type: string
          example: smtp.gmail.com
        port:
          type: integer
          default: 587
        secure:
          type: boolean
          default: false
        user:
          type: string
        pass:
          type: string
          format: password
        fromEmail:
          type: string
          format: email
        fromName:
          type: string
        isDefault:
          type: boolean
      required: [host, user, pass, fromEmail]
    SMTPConfigUpdate:
      type: object
      properties:
        name:
          type: string
        host:
          type: string
        port:
          type: integer
        secure:
          type: boolean
        user:
          type: string
        pass:
          type: string
          format: password
        fromEmail:
          type: string
          format: email
        fromName:
          type: string
        isDefault:
          type: boolean
    SMTPTestRequest:
      type: object
      properties:
        host:
          type: string
        port:
          type: integer
          default: 587
        secure:
          type: boolean
        user:
          type: string
        pass:
          type: string
          format: password
      required: [host, user, pass]

    # -- Campaigns --
...
security:
  - cookieAuth: []
  - apiKeyAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /config/smtp:
    get:
      tags: [Config]
      summary: Get SMTP configurations (legacy)
      operationId: getSmtpConfigs
      responses:
        '200':
          description: SMTP configurations with default indicator.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

    post:
      tags: [Config]
      summary: Create SMTP configuration
      operationId: createSmtpConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SMTPConfigCreate'
      responses:
        '200':
          description: Configuration created.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          configId:
                            type: string
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /config/smtp/{configId}:
    put:
      tags: [Config]
      summary: Update SMTP configuration
      operationId: updateSmtpConfig
      parameters:
        - name: configId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SMTPConfigUpdate'
      responses:
        '200':
          description: Configuration updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Configuration not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [Config]
      summary: Delete SMTP configuration
      operationId: deleteSmtpConfig
      parameters:
        - name: configId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Configuration deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Configuration not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /config/smtp/{configId}/default:
    post:
      tags: [Config]
      summary: Set a configuration as default
      operationId: setDefaultConfig
      parameters:
        - name: configId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Default configuration updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Configuration not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /config/smtp/test:
    post:
      tags: [Config]
      summary: Test SMTP connection with ad-hoc credentials
      operationId: testSmtpConnection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SMTPTestRequest'
      responses:
        '200':
          description: Connection test result.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          valid:
                            type: boolean
  /config/test/{configId}:
    post:
      tags: [Config]
      summary: Test a saved SMTP configuration
      operationId: testSavedConfig
      parameters:
        - name: configId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Connection test result.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          valid:
                            type: boolean
        '404':
          description: Configuration not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # =========================================================================
  # Send
  # =========================================================================
...